+++ lib/error_catch_boundary.dart
@@ -1,3 +1,6 @@
+export 'src/auto_retry_config.dart';
 export 'src/default_error_widget.dart';
 export 'src/error_boundary.dart';
+export 'src/error_boundary_controller.dart';
 export 'src/error_details.dart';
+export 'src/global_error_boundary_config.dart';